Client code just merrily uses a morph::VisualModel<int>
